requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
# Screenshot options (at least one is required)
pyautogui>=0.9.53
mss>=6.1.0
# Optional: for keyboard shortcuts
keyboard>=0.13.5
# For Azure OpenAI integration
openai>=1.0.0
python-dotenv>=1.0.0
azure-ai-inference